+++ src/util/router.ts
@@ -17,3 +17,14 @@ export const getCanonicalURL = (title: string): string => {
     const removedSpaces = cleaned.split(" ").join("-");
     return removedSpaces;
 };
+
+export const getBaseURL = (): string => {
+    const baseUrl = process.env.LOCAL
+        ? "http://localhost:3000"
+        : {
+              test: "https://alan-john-portfolio.vercel.app",
+              development: "https://alan-john-portfolio.vercel.app",
+              production: "https://www.alanjohn.dev.dev",
+          }[process.env.NODE_ENV];
+    return baseUrl;
+};