+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2022-02-19
+### Added
+- Function to access BaseURL
+- Added `CustomArticleJSONld` tag for rich text search
+
 ## [1.0.1] - 2022-02-19
 ### Added 
 - Added Sitemap and robots.txt